src/main/java/com/asde/junit/Flight.java
package com.asde.junit;

import java.util.HashSet;
import java.util.Set;
import java.util.regex.Matcher;
import java.util.regex.Pattern;

public class Flight {
    private String flightNumber;
    private int seats;
    private Set<Passenger> passengers = new HashSet<>();
    private static String flightNumberRegex = "^[A-Z]{2}\\d{3,4}$";
    private static Pattern pattern = Pattern.compile(flightNumberRegex);

    public Flight(String flightNumber, int seats) {
        Matcher matcher = pattern.matcher(flightNumber);
        if(!matcher.matches()) {
            throw new RuntimeException("Invalid Flight Number");
        }
        this.flightNumber = flightNumber;
        this.seats = seats;
    } // constructor

    public boolean addPassenger(Passenger passenger) {
        if (getPassengersSize() >= seats) {
            throw new RuntimeException("Not enough seats for flight "
                    + getFlightNumber());
        }
        passenger.setFlight(this);
        return passengers.add(passenger);
    }

    public boolean removePassenger(Passenger passenger) {
        return passengers.remove(passenger);
    }

    public int getPassengersSize() {
        return passengers.size();
    }

    public String getFlightNumber() {
        return flightNumber;
    }

    public void setFlightNumber(String flightNumber) {
        this.flightNumber = flightNumber;
    }

    public int getSeats() {
        return seats;
    }

    public void setSeats(int seats) {
        this.seats = seats;
    }

    public Set<Passenger> getPassengers() {
        return passengers;
    }

    public void setPassengers(Set<Passenger> passengers) {
        this.passengers = passengers;
    }
} // end class
